Code style and legibility
