+++ package.json
@@ -42,6 +42,7 @@
 		"comlink": "^4.4.1",
 		"compressible": "2.0.18",
 		"compression": "1.7.4",
+		"copy-webpack-plugin": "11.0.0",
 		"express-fileupload": "1.4.0",
 		"file-saver": "^2.0.5",
 		"follow-redirects": "1.15.2",
