+++ packages/wordpress-playground-block/src/view.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from '@wordpress/element';
-import PlaygroundDemo from './components/PlaygroundDemo';
+import PlaygroundPreview from './components/playground-preview';
 
 const playgroundDemo = Array.from(
 	document.getElementsByClassName('wordpress-playground-block')
@@ -13,5 +13,5 @@ for (const element of playgroundDemo) {
 		atob(rootElement.dataset['attributes'] || '')
 	);
 
-	root.render(<PlaygroundDemo {...attributes} />);
+	root.render(<PlaygroundPreview {...attributes} />);
 }
